dom/tests/mochitest/bugs/test_bug317448.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=317448
-->
<head>
  <title>Test for Bug 317448</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=317448">Mozilla Bug 317448</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 317448 **/
var x = new XMLHttpRequest();
x.open("GET", document.location.href);
x.send("");
is(x instanceof SpecialPowers.Ci.nsIInterfaceRequestor,
   true, "Must be interface requestor");

var count = {};
var interfaces = SpecialPowers.wrap(x).
                 QueryInterface(SpecialPowers.Ci.nsIClassInfo).
                 getInterfaces(count).
                 map(function(id) {
                            id = SpecialPowers.wrap(id);
                            return SpecialPowers.wrap(SpecialPowers.Components).interfacesByID[id].toString();
                     });
isnot(interfaces.indexOf("nsIInterfaceRequestor"), "-1",
      "Must have interface requestor classinfo");
</script>
</pre>
</body>
</html>
